move font : to text
