+++ src/components/text/RawText.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components/native';
-import { Font } from 'src/draw/font';
+import { Font } from 'src/draw/text/font';
 
 const RawText = styled.Text<Font>`
   font-family: ${({ fontFamily }: Font) => fontFamily};
